絵画未来の絵が消えるようになりました

+++ Solution/App/Future.h
@@ -14,7 +14,7 @@ public:
 	GRAPH getGraph( ) const;
 	void update( );
 	void change( );
-	void erase( Vector pos, int radius ); 
+	void erase( Vector pos, double radius ); 
 private:
 	void load( );
 private:
